requirements.txt
Flask==2.3.2
Flask_SocketIO==5.3.6
matplotlib==3.8.2
numpy==2.0.0
pyserial==3.5
scipy==1.14.0
